Factor out `FqVector::is_zero`
