+++ ext/crates/fp/src/vector/inner.rs
@@ -123,6 +123,30 @@ impl<const A: bool, R: Repr, F: Field> FqVectorBase<A, R, F> {
         )
     }
 
+    pub fn is_zero(&self) -> bool {
+        if A {
+            return self.limbs().iter().all(|&x| x == 0);
+        }
+
+        let limb_range = self.limb_range();
+        if limb_range.is_empty() {
+            return true;
+        }
+        let (min_mask, max_mask) = self.limb_masks();
+        if self.limbs()[limb_range.start] & min_mask != 0 {
+            return false;
+        }
+
+        let inner_range = self.limb_range_inner();
+        if !inner_range.is_empty() && self.limbs()[inner_range].iter().any(|&x| x != 0) {
+            return false;
+        }
+        if self.limbs()[limb_range.end - 1] & max_mask != 0 {
+            return false;
+        }
+        true
+    }
+
     pub fn entry(&self, index: usize) -> FieldElement<F> {
         debug_assert!(
             index < self.len(),
